Added some steps to basic.t to test that git-cvs pull pulls the changes to the right branches of git.
Noticed that my earlier de-crufting of unneeded declarations had
removed a line setting $PATH to ensure this dist's version of git-cvs
is invoked.  Added a exe_map parameter to the MyTest::Replay module to
allow git-cvs to be explicitly defined as the dist's executable.

Removed the commmented -i flag from the git-cvsimport command
invokation.  Uncommenting it makes git-cvs fail to import the initial
state properly in basic.t.